requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# LangChain + OpenAI
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
langsmith>=0.1.70
openai>=1.30.0

# SSE streaming support
sse-starlette>=1.8.2

# Data processing
pandas>=2.0.0
numpy>=1.26.0
